requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.66.0
anyio==4.10.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
async-lru==2.0.5
asyncer==0.0.8
attrs==25.3.0
babel==2.17.0
backoff==2.2.1
beautifulsoup4==4.13.5
bidict==0.23.1
bleach==6.2.0
certifi==2025.8.3
cffi==2.0.0
chainlit==2.7.2
charset-normalizer==3.4.3
chevron==0.14.0
click==8.2.1
colorama==0.4.6
contourpy
cryptography==45.0.7
cuid==0.4
cycler==0.12.1
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.18
distro==1.9.0
emoji==2.14.1
faiss-cpu==1.12.0
fastapi==0.116.1
fastjsonschema==2.21.2
filelock==3.19.1
filetype==1.2.0
fonttools==4.59.2
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
googleapis-common-protos==1.70.0
gritql==0.2.0
grpcio==1.74.0
h11==0.16.0
hf-xet==1.1.9
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
idna==3.10
inflection==0.5.1
ipywidgets==8.1.7
isoduration==20.11.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
json5==0.12.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.7
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kaggle==1.7.4.5
kagglehub==0.3.13
kiwisolver==1.4.9
langchain==0.3.27
langchain-cli==0.0.37
langchain-community==0.3.29
langchain-core==0.3.75
langchain-huggingface==0.3.1
langchain-openai==0.3.32
langchain-text-splitters==0.3.11
langdetect==1.0.9
langserve==0.3.1
langsmith==0.4.27
lark==1.2.2
Lazify==0.4.0
literalai==0.1.201
lxml==6.0.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.6
mcp==1.13.1
mdurl==0.1.2
mistune==3.1.4
monotonic==1.6
mpmath==1.3.0
multidict==6.6.4
mypy_extensions==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
networkx
nltk==3.9.1
notebook==7.4.5
notebook_shim==0.2.4
numpy
olefile==0.47
openai==1.107.0
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-exporter-otlp-proto-http==1.36.0
opentelemetry-instrumentation==0.57b0
opentelemetry-instrumentation-alephalpha==0.46.2
opentelemetry-instrumentation-anthropic==0.46.2
opentelemetry-instrumentation-bedrock==0.46.2
opentelemetry-instrumentation-chromadb==0.46.2
opentelemetry-instrumentation-cohere==0.46.2
opentelemetry-instrumentation-crewai==0.46.2
opentelemetry-instrumentation-google-generativeai==0.46.2
opentelemetry-instrumentation-groq==0.46.2
opentelemetry-instrumentation-haystack==0.46.2
opentelemetry-instrumentation-lancedb==0.46.2
opentelemetry-instrumentation-langchain==0.46.2
opentelemetry-instrumentation-llamaindex==0.46.2
opentelemetry-instrumentation-logging==0.57b0
opentelemetry-instrumentation-marqo==0.46.2
opentelemetry-instrumentation-mcp==0.46.2
opentelemetry-instrumentation-milvus==0.46.2
opentelemetry-instrumentation-mistralai==0.46.2
opentelemetry-instrumentation-ollama==0.46.2
opentelemetry-instrumentation-openai==0.46.2
opentelemetry-instrumentation-openai-agents==0.46.2
opentelemetry-instrumentation-pinecone==0.46.2
opentelemetry-instrumentation-qdrant==0.46.2
opentelemetry-instrumentation-redis==0.57b0
opentelemetry-instrumentation-replicate==0.46.2
opentelemetry-instrumentation-requests==0.57b0
opentelemetry-instrumentation-sagemaker==0.46.2
opentelemetry-instrumentation-sqlalchemy==0.57b0
opentelemetry-instrumentation-threading==0.57b0
opentelemetry-instrumentation-together==0.46.2
opentelemetry-instrumentation-transformers==0.46.2
opentelemetry-instrumentation-urllib3==0.57b0
opentelemetry-instrumentation-vertexai==0.46.2
opentelemetry-instrumentation-watsonx==0.46.2
opentelemetry-instrumentation-weaviate==0.46.2
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
opentelemetry-semantic-conventions-ai==0.4.13
opentelemetry-util-http==0.57b0
orjson==3.11.3
overrides==7.7.0
pandas==2.3.2
pandocfilters==1.5.1
pillow==11.3.0
posthog==3.25.0
prometheus_client==0.22.1
propcache==0.3.2
protobuf==6.32.0
pycparser==2.23
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.4
pypdf==6.0.0
python-dotenv==1.1.1
python-engineio==4.12.2
python-iso639==2025.2.18
python-json-logger==3.3.0
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
python-slugify==8.0.4
python-socketio==5.13.0
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.14.1
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.1.0
rpds-py==0.27.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy
Send2Trash==1.8.3
sentence-transformers==5.1.0
sentencepiece==0.2.1
sentry-sdk==2.37.1
shellingham==1.5.4
simple-websocket==1.1.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
sse-starlette==1.8.2
starlette==0.47.3
sympy==1.14.0
syncer==2.0.3
tenacity==9.1.2
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.6.0
tiktoken==0.11.0
tinycss2==1.4.0
tokenizers==0.22.0
tomli==2.2.1
tomlkit==0.13.3
torch==2.8.0
tqdm==4.67.1
traceloop-sdk==0.46.2
transformers==4.56.1
typer==0.17.4
types-python-dateutil==2.9.0.20250822
typing-inspect==0.9.0
typing-inspection==0.4.1
tzdata==2025.2
unstructured==0.18.14
unstructured-client==0.42.3
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.35.0
wandb==0.21.3
watchfiles==0.24.0
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.14
wrapt==1.17.3
wsproto==1.2.0
yarl==1.20.1
zstandard==0.24.0
datasets
